% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_based_algos.R
\name{rule_based_algos}
\alias{rule_based_algos}
\title{Apply rule-based algorithms}
\usage{
rule_based_algos(dat, dx1, algorithm, ...)
}
\arguments{
\item{dat}{A dataframe with observations to be classified.}

\item{dx1}{The primary diagnosis, coded in ICD-10. Include only the first three
characters of the ICD-10 code.}

\item{algorithm}{Selection of rule-based algorithms to apply to the data. Choices
include "billings", "ahrq", "weinick", "aihw", "sundmacher", "nicholl", and
"scheuttig".}

\item{dx2}{A secondary diagnosis, if available, also coded in ICD-10. Include
only the first three characters of the code.}

\item{age}{Patient age.}
}
\value{
The dataframe \code{dat} with additional columns appended for each selected
algorithm. Appended columns consist of TRUE/FALSE indicators, with TRUE indicating
that the observation is classified as low acuity by the selected algorithm.
}
\description{
This function applies original ICD code rule-based algorithms to your data. Seven
algorithms available in ICD-10 native coding are provided for use. Diagnosis codes
are coded using ICD-10 and included at the “category” level, corresponding to
the first three characters of the codes. When indicated, this function follows
the conditional logic of algorithms provided.
}
\examples{
nhamcs <- rule_based_algos(nhamcs, "icd10", dx1 = "diag1", 
     algorithm = c("billings", "ahrq", "weinick", "aihw", "sundmacher", 
     "nicholl", "scheuttig"), dx2 ="diag2", age = "age")

}
